chore: abstract out sorting functions
