+++ src/routes/received.tsx
@@ -7,6 +7,7 @@ import Table from '../components/Table';
 import { TableLoader } from '../components/TableLoader';
 import type { Submission } from '../components/shared/types';
 import { forNewSubmissions } from '../queries';
+import { dateStringSortingFn, mrrcCellText, nullableBooleanSortingFn } from '../sortingFns';
 
 const columnHelper = createColumnHelper<Submission>();
 const columns = [
@@ -34,43 +35,13 @@ const columns = [
   columnHelper.accessor('date', {
     id: 'date',
     header: () => 'Submission Date',
-    sortingFn: (rowA, rowB, columnId) => {
-      const parse = (value: string) => {
-        const ms = Date.parse(value);
-        return Number.isNaN(ms) ? null : ms;
-      };
-
-      const timeA = parse(rowA.getValue<string>(columnId));
-      const timeB = parse(rowB.getValue<string>(columnId));
-
-      if (timeA === null && timeB === null) return 0;
-      if (timeA === null) return -1;
-      if (timeB === null) return 1;
-      return timeA - timeB;
-    },
+    sortingFn: dateStringSortingFn,
   }),
   columnHelper.accessor('mrrc', {
     id: 'mrrc',
     header: () => 'MRRC',
-    cell: (info) => {
-      const value = info.getValue();
-      if (value === undefined) {
-        return 'Unknown';
-      }
-
-      return value ? 'Yep' : 'Nope';
-    },
-    sortingFn: (rowA, rowB, columnId) => {
-      const rank = (value: boolean | undefined) => {
-        if (value === undefined) {
-          return 0;
-        }
-
-        return value ? 2 : 1;
-      };
-
-      return rank(rowA.getValue<boolean | undefined>(columnId)) - rank(rowB.getValue<boolean | undefined>(columnId));
-    },
+    cell: (info) => mrrcCellText(info.getValue()),
+    sortingFn: nullableBooleanSortingFn,
   }),
 ];
