+++ src/routes/rejected.tsx
@@ -7,6 +7,7 @@ import Table from '../components/Table';
 import { TableLoader } from '../components/TableLoader';
 import type { RejectedSubmission } from '../components/shared/types';
 import { forRejectedSubmissions } from '../queries';
+import { dateStringSortingFn, localeStringSortingFn } from '../sortingFns';
 
 const columnHelper = createColumnHelper<RejectedSubmission>();
 
@@ -39,30 +40,13 @@ export default function Rejected() {
       columnHelper.accessor('date', {
         id: 'date',
         header: () => 'Rejected Date',
-        sortingFn: (rowA, rowB, columnId) => {
-          const parse = (value: string) => {
-            const ms = Date.parse(value);
-            return Number.isNaN(ms) ? null : ms;
-          };
-
-          const timeA = parse(rowA.getValue<string>(columnId));
-          const timeB = parse(rowB.getValue<string>(columnId));
-
-          if (timeA === null && timeB === null) return 0;
-          if (timeA === null) return -1;
-          if (timeB === null) return 1;
-          return timeA - timeB;
-        },
+        sortingFn: dateStringSortingFn,
       }),
       columnHelper.accessor('rejectedFrom', {
         id: 'rejectedFrom',
         header: () => 'Rejected by',
         cell: (info) => <span aria-label={`Rejected by ${info.row.original.rejectedBy}`}>{info.getValue()}</span>,
-        sortingFn: (rowA, rowB, columnId) => {
-          const valueA = rowA.getValue<string>(columnId);
-          const valueB = rowB.getValue<string>(columnId);
-          return valueA.localeCompare(valueB);
-        },
+        sortingFn: localeStringSortingFn,
       }),
       columnHelper.accessor('reason', {
         id: 'reason',
